Unfold the model/webbapp duality
Toying at work with JRebel and its poor cousing Spring Loaded, I came to
realize that the current project setup won't work with Spring Loaded
thanks to how Gradle handles subproject dependencies (Spring Loaded
doesn't instrument JARs). Searching for a solution, I found out that
separating application layers into subprojects is highly disregarded by
a lot of people (not sure if enough to qualify it as an antipattern,
though).
Eg. http://www.buzdin.lv/2012/08/maven-modules-considered-harmful.html
That led me to the conclusion that I should rework the setup, not
because Spring Loaded, but as a matter of exercise. I never really liked
the idea of having a subproject just with model classes, but it was the
easiest way to deal with EclipseLink static weaving when I had no
fucking idea of how Gradle worked. I still don't have it, but at least
now I can hack through it. There's some work to do on the weaving task
to make it more compatible with Gradle's up-to-date checking, and then
make it a Gradle plugin. But that is another story...